pyproject.toml
[project]
name = "dumbledore"
version = "0.1.0"
description = ""
authors = [
    {name = "dev-Caprisunn",email = "ikpebaroselyn235@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "djoser (>=2.3.3,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
